attestation: refactoring/renaming
